backend/apm-ai-service/python-sidecar/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.28.3
numpy>=2.2.0
scikit-learn>=1.5.2
pandas>=2.2.3
openai>=1.82.0
anthropic>=0.52.0
